requirements.txt
psycopg2-binary
rich
pytest
sqlalchemy
pandas
python-dotenv
